+++ .cspell.json
@@ -4,7 +4,7 @@
     "words": [
         "zenfuse",
         "zefu",
-        "bithumb",
+        "bitglobal",
         "deepmerge",
         "bitglobal",
         "binance",
