+++ .env.example
@@ -7,8 +7,8 @@ FTX_PUBLIC_KEY=
 FTX_SECRET_KEY=
 
 # Create new keys on https://www.bitglobal.com/en-us/account/user/api-key/list
-BITHUMB_PUBLIC_KEY=
-BITHUMB_SECRET_KEY=
+BITGLOBAL_PUBLIC_KEY=
+BITGLOBAL_SECRET_KEY=
 
 # Create new keys on https://www.okx.com/ru/account/my-api
 OKX_PUBLIC_KEY=
