+++ tests/exchanges/bitglobal/bitglobal.test.js
@@ -4,18 +4,18 @@ const masterTest = require('../../master.test');
 const createScope = require('./scope');
 const checkProcessHasVariables = require('../../helpers/validateEnv');
 const createEnv = require('../../helpers/createEnv');
-const BitglobalApiError = require('../../../src/exchanges/bithumb/errors/api.error');
+const BitglobalApiError = require('../../../src/exchanges/bitglobal/errors/api.error');
 
 if (isEnd2EndTest) {
-    checkProcessHasVariables(['BITHUMB_PUBLIC_KEY', 'BITHUMB_SECRET_KEY']);
+    checkProcessHasVariables(['BITGLOBAL_PUBLIC_KEY', 'BITGLOBAL_SECRET_KEY']);
 }
 
 /**
  * @type {import('../../master.test').MasterTestEnvironment}
  */
 const env = createEnv({
-    API_PUBLIC_KEY: process.env.BITHUMB_PUBLIC_KEY,
-    API_PRIVATE_KEY: process.env.BITHUMB_SECRET_KEY,
+    API_PUBLIC_KEY: process.env.BITGLOBAL_PUBLIC_KEY,
+    API_PRIVATE_KEY: process.env.BITGLOBAL_SECRET_KEY,
     NOT_EXECUTABLE_ORDER: {
         symbol: 'BTC/USDT',
         type: 'limit',
